+++ src/Kentico.Kontent.ModelGenerator.Core/Generators/Class/ClassCodeGenerator.cs
@@ -1,10 +1,12 @@
 ﻿using System;
+using Kentico.Kontent.ModelGenerator.Core.Common;
+using Kentico.Kontent.ModelGenerator.Core.Helpers;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Formatting;
 
-namespace Kentico.Kontent.ModelGenerator.Core
+namespace Kentico.Kontent.ModelGenerator.Core.Generators.Class
 {
     public abstract class ClassCodeGenerator
     {
