+++ src/Kentico.Kontent.ModelGenerator.Core/Generators/Class/ManagementClassCodeGenerator.cs
@@ -1,10 +1,11 @@
 ﻿using System.Linq;
+using Kentico.Kontent.Management.Modules.ModelBuilders;
+using Kentico.Kontent.ModelGenerator.Core.Common;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Kentico.Kontent.Management.Modules.ModelBuilders;
 using Newtonsoft.Json.Serialization;
 
-namespace Kentico.Kontent.ModelGenerator.Core
+namespace Kentico.Kontent.ModelGenerator.Core.Generators.Class
 {
     public class ManagementClassCodeGenerator : ClassCodeGenerator
     {
