+++ src/Kentico.Kontent.ModelGenerator.Core/Generators/Class/PartialClassCodeGenerator.cs
@@ -1,8 +1,9 @@
 ﻿using System;
+using Kentico.Kontent.ModelGenerator.Core.Common;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace Kentico.Kontent.ModelGenerator.Core
+namespace Kentico.Kontent.ModelGenerator.Core.Generators.Class
 {
     public class PartialClassCodeGenerator : DeliveryClassCodeGeneratorBase
     {
