+++ src/Kentico.Kontent.ModelGenerator.Core/Helpers/ElementIdHelper.cs
@@ -6,7 +6,7 @@ using Kentico.Kontent.Management.Models.Types;
 using Kentico.Kontent.Management.Models.Types.Elements;
 using Kentico.Kontent.Management.Models.TypeSnippets;
 
-namespace Kentico.Kontent.ModelGenerator.Core
+namespace Kentico.Kontent.ModelGenerator.Core.Helpers
 {
     public static class ElementIdHelper
     {
