+++ test/Kentico.Kontent.ModelGenerator.Tests/ClassCodeGeneratorFactoryTests.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using Kentico.Kontent.ModelGenerator.Core;
+using Kentico.Kontent.ModelGenerator.Core.Common;
 using Kentico.Kontent.ModelGenerator.Core.Configuration;
+using Kentico.Kontent.ModelGenerator.Core.Generators.Class;
 using Xunit;
 
 namespace Kentico.Kontent.ModelGenerator.Tests
