+++ test/Kentico.Kontent.ModelGenerator.Tests/ClassCodeGeneratorTests.cs
@@ -1,7 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Reflection;
-using Kentico.Kontent.ModelGenerator.Core;
+using Kentico.Kontent.ModelGenerator.Core.Common;
+using Kentico.Kontent.ModelGenerator.Core.Generators.Class;
 using Xunit;
 
 namespace Kentico.Kontent.ModelGenerator.Tests
