+++ test/Kentico.Kontent.ModelGenerator.Tests/ClassDefinitionTests.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using Kentico.Kontent.Delivery.Abstractions;
 using Kentico.Kontent.ModelGenerator.Core;
+using Kentico.Kontent.ModelGenerator.Core.Common;
 using Moq;
 using Xunit;
