+++ test/Kentico.Kontent.ModelGenerator.Tests/ElementIdHelperTests.cs
@@ -5,6 +5,7 @@ using Kentico.Kontent.Management.Models.Types;
 using Kentico.Kontent.Management.Models.Types.Elements;
 using Kentico.Kontent.Management.Models.TypeSnippets;
 using Kentico.Kontent.ModelGenerator.Core;
+using Kentico.Kontent.ModelGenerator.Core.Helpers;
 using Newtonsoft.Json.Linq;
 using Xunit;
