+++ test/Kentico.Kontent.ModelGenerator.Tests/PartialClassCodeGeneratorTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.IO;
-using Kentico.Kontent.ModelGenerator.Core;
+using Kentico.Kontent.ModelGenerator.Core.Common;
+using Kentico.Kontent.ModelGenerator.Core.Generators.Class;
 using Xunit;
 
 namespace Kentico.Kontent.ModelGenerator.Tests
