+++ test/Kentico.Kontent.ModelGenerator.Tests/TextHelpersTests.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using Kentico.Kontent.ModelGenerator.Core;
+using Kentico.Kontent.ModelGenerator.Core.Common;
+using Kentico.Kontent.ModelGenerator.Core.Helpers;
 using Xunit;
 
 namespace Kentico.Kontent.ModelGenerator.Tests
